third_party/blink/web_tests/platform/win/virtual/text-antialias/hyphens/hyphens-parsing-001-expected.txt

This is a testharness.js-based test.
PASS Initial value
FAIL hyphens: auto assert_equals: expected "auto" but got "manual"
PASS hyphens: manual
PASS hyphens: none
FAIL hyphens should inherit assert_equals: expected "auto" but got "manual"
Harness: the test ran to completion.
